Move literature related components to a new literatures folder
