+++ app/packs/src/components/elements/DetailsTabLiteratures.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { ListGroup, ListGroupItem, Button, Row, Col } from 'react-bootstrap';
 import uuid from 'uuid';
 import Immutable from 'immutable';
-import { Citation, doiValid, sanitizeDoi, groupByCitation, AddButton, LiteratureInput, LiteralType } from 'src/components/LiteratureCommon';
+import { Citation, doiValid, sanitizeDoi, groupByCitation, AddButton, LiteratureInput, LiteralType } from 'src/components/literature/LiteratureCommon';
 import Sample from 'src/models/Sample';
 import Reaction from 'src/models/Reaction';
 import ResearchPlan from 'src/models/ResearchPlan';
@@ -13,8 +13,8 @@ import LiteraturesFetcher from 'src/fetchers/LiteraturesFetcher';
 import UserStore from 'src/stores/alt/stores/UserStore';
 import NotificationActions from 'src/stores/alt/actions/NotificationActions';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
-import CitationPanel from 'src/components/CitationPanel';
-import { CitationTypeMap } from 'src/components/CitationType';
+import CitationPanel from 'src/components/literature/CitationPanel';
+import { CitationTypeMap } from 'src/components/literature/CitationType';
 
 const Cite = require('citation-js');
 require('@citation-js/plugin-isbn');
