+++ app/packs/src/components/literature/CitationType.js
